♻️ adicionada camada intermediária entre a biblioteca e a API
