+++ src/services/whatsapp/index.js
@@ -1,4 +1,3 @@
 const { buildClient } = require('./client');
-const { sendMessage } = require('./send');
 
-module.exports = { sendMessage, buildClient };
+module.exports = { buildClient };
